oops-- avoid circular import
